Function recipeText, renders condtionally steps
